+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## v0.1.2
 - Added support for installing Python on most Linux distros
-- Fixed a bug related to creating `pyflow` directory
-- Fixed a bug in specifying package url with the `publish` command.
 - Wheel is now installed directly, instead of with Pip; should only be dependent on
 pip now to install `twine`.
+- Now doesn't ask to choose between aliases pointing to the same Python install.
+- Fixed a bug related to creating `pyflow` directory
+- Fixed a bug in specifying package url with the `publish` command.
+
 
 ## v0.1.1
 - Fixed a bug, where spaces could prevent console scripts from being installed